# openapi/dist/openapi.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: Reel Fake Api
  description: A REST api for buying or renting a movie dvd (only for educational purposes)
  license:
    name: MIT License
    url: https://mit-license.org
  contact:
    name: Pratap Reddy
    email: pratapreddy15@gmail.com
servers:
  - url: http://reelfake.cloud/api
    description: Production server
  - url: http://127.0.0.1:8000/api
    description: Local server
security:
  - cookieAuth: []
paths:
...
  /staff/{id}/forgot_password:
    put:
      summary: Change the staff password which is forgotten
      operationId: forgotStaffPassword
      tags:
        - staff
      parameters:
        - in: path
          name: id
          required: true
          description: The id of the staff
          schema:
            type: integer
      requestBody:
        description: Payload for the new password
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newPassword:
                  type: string
                  format: password
                confirmedNewPassword:
                  type: string
                  format: password
      responses:
        '200':
          description: Successfully changed the password
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: number
                  email:
                    type: string
                    format: email
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
  /staff/{id}/change_password:
    put:
      summary: Change the staff password
      operationId: changeStaffPassword
      tags:
        - staff
      parameters:
        - in: path
          name: id
          required: true
          description: Staff id
          schema:
            type: integer
      requestBody:
        description: Payload for the current and new password
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                  format: password
                newPassword:
                  type: string
                  format: password
                confirmedNewPassword:
                  type: string
                  format: password
      responses:
        '200':
          description: Successfully changed the password
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: number
                  email:
                    type: string
                    format: email
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
  /staff:
    get:
      summary: Get list of all staff (optionally filter by residential state and city)
      operationId: getStaff
      tags:
        - staff
      parameters:
        - in: query
          name: page
          required: false
          description: |
            The page number (which starts with 1) to fetch the set of staff from.
            The staff are split into multiple pages with each page having a limited number of staff.
          schema:
            type: integer
            default: 1
        - in: query
          name: first_name
          required: false
          description: |
            The first name of the staff.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: first_name=%tap, first_name=pra% and first_name=%ata%.
          schema:
            type: string
        - in: query
          name: last_name
          required: false
          description: |
            The last name of the staff.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: last_name=%tap, last_name=pra% and last_name=%ata%.
          schema:
            type: string
        - in: query
          name: email
          required: false
          description: |
            The email of the staff.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: email=%tap, email=pra% and email=%ata%.
          schema:
            type: string
            format: email
        - in: query
          name: phone_number
          required: false
          description: |
            The phone number of the staff.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: phone_number=%456, phone_number=123% and phone_number=%456%.
          schema:
            type: string
        - in: query
          name: address_line
          required: false
          description: |
            The address line of the staff resident.
            Filters the staff by the address line of the staff address.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: address_line=%street, address_line=any% and address_line=%str%.
          schema:
            type: string
        - in: query
          name: city
          required: false
          description: |
            The city of the staff resident.
            Filters the staff by the city of the staff resident.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: city=%abad, city=hyder% and city=%aba%.
          schema:
            type: string
        - in: query
          name: state
          required: false
          description: |
            The state of the staff resident.
            Filters the staff by the state of the staff resident.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: state=%gana, state=tel% and state=%anga%.
          schema:
            type: string
        - in: query
          name: country
          required: false
          description: |
            The country of the staff resident.
            Filters the staff by the country of the staff resident.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: country=%dia, country=ind% and country=%ndi%.
          schema:
            type: string
        - in: query
          name: postal_code
          required: false
          description: |
            The postal_code of the staff resident.
            Filters the staff by the postal_code of the staff resident.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: postal_code=%048, postal_code=500% and postal_code=%000%.
          schema:
            type: string
      responses:
        '200':
          description: Successfully returned list of all staff
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/staffSummary'
                  length:
                    type: integer
                  pagination:
                    $ref: '#/components/schemas/pagination'
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '500':
          $ref: '#/components/responses/internalServerError'
    post:
      summary: Add staff member
      operationId: createStaff
      tags:
        - staff
      requestBody:
        description: Payload for the new staff
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                phoneNumber:
                  type: string
                avatar:
                  type: string
                  format: uri
                storeId:
                  type: integer
                  format: number
                address:
                  type: object
                  properties:
                    addressLine:
                      type: string
                    cityName:
                      type: string
                    stateName:
                      type: string
                    postalCode:
                      type: string
      responses:
        '201':
          description: Successfully created staff
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: number
                    readOnly: true
                  firstName:
                    type: string
                  lastName:
                    type: string
                  email:
                    type: string
                    format: email
                  active:
                    type: boolean
                  phoneNumber:
                    type: string
                  avatar:
                    type: string
                    format: uri
                  storeId:
                    type: integer
                    format: number
                  address:
                    $ref: '#/components/schemas/address'
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '500':
          $ref: '#/components/responses/internalServerError'
  /staff/summary:
    get:
      summary: Get list of all staff (optionally filter by residential state and city)
      operationId: getStaffSummary
      tags:
        - staff
      parameters:
        - in: query
          name: page
          required: false
          description: |
            The page number (which starts with 1) to fetch the set of staff from.
            The staff are split into multiple pages with each page having a limited number of staff.
          schema:
            type: integer
            default: 1
        - in: query
          name: first_name
          required: false
          description: |
            The first name of the staff.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: first_name=%tap, first_name=pra% and first_name=%ata%.
          schema:
            type: string
        - in: query
          name: last_name
          required: false
          description: |
            The last name of the staff.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: last_name=%tap, last_name=pra% and last_name=%ata%.
          schema:
            type: string
        - in: query
          name: email
          required: false
          description: |
            The email of the staff.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: email=%tap, email=pra% and email=%ata%.
          schema:
            type: string
            format: email
        - in: query
          name: phone_number
          required: false
          description: |
            The phone number of the staff.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: phone_number=%456, phone_number=123% and phone_number=%456%.
          schema:
            type: string
        - in: query
          name: address_line
          required: false
          description: |
            The address line of the staff resident.
            Filters the staff by the address line of the staff address.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: address_line=%street, address_line=any% and address_line=%str%.
          schema:
            type: string
        - in: query
          name: city
          required: false
          description: |
            The city of the staff resident.
            Filters the staff by the city of the staff resident.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: city=%abad, city=hyder% and city=%aba%.
          schema:
            type: string
        - in: query
          name: state
          required: false
          description: |
            The state of the staff resident.
            Filters the staff by the state of the staff resident.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: state=%gana, state=tel% and state=%anga%.
          schema:
            type: string
        - in: query
          name: country
          required: false
          description: |
            The country of the staff resident.
            Filters the staff by the country of the staff resident.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: country=%dia, country=ind% and country=%ndi%.
          schema:
            type: string
        - in: query
          name: postal_code
          required: false
          description: |
            The postal_code of the staff resident.
            Filters the staff by the postal_code of the staff resident.
            Prefix % for starts with search, suffix % for ends with search and enclose with % for contains search.
            Example: postal_code=%048, postal_code=500% and postal_code=%000%.
          schema:
            type: string
      responses:
        '200':
          description: Successfully returned list of all staff
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          format: number
                        firstName:
                          type: string
                        lastName:
                          type: string
                        email:
                          type: string
                          format: email
                        active:
                          type: boolean
                  length:
                    type: integer
                  pagination:
                    $ref: '#/components/schemas/pagination'
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '500':
          $ref: '#/components/responses/internalServerError'
  /staff/managers:
    get:
      summary: Get list of store managers
      operationId: getStoreManagers
      tags:
        - staff
      responses:
        '200':
          description: Successfully returned list of store managers
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          format: number
                        firstName:
                          type: string
                        lastName:
                          type: string
                        email:
                          type: string
                          format: email
                        active:
                          type: boolean
                        phoneNumber:
                          type: string
                        avatar:
                          type: string
                        address:
                          $ref: '#/components/schemas/address'
                        store:
                          type: object
                          properties:
                            id:
                              type: integer
                              format: number
                            phoneNumber:
                              type: string
                            address:
                              $ref: '#/components/schemas/address'
                  length:
                    type: integer
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '500':
          $ref: '#/components/responses/internalServerError'
  /staff/managers/summary:
    get:
      summary: Get list of store managers
      operationId: getStoreManagersSummary
      tags:
        - staff
      responses:
        '200':
          description: Successfully returned list of store managers
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          format: number
                        firstName:
                          type: string
                        lastName:
                          type: string
                        email:
                          type: string
                          format: email
                        active:
                          type: boolean
                  length:
                    type: integer
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '500':
          $ref: '#/components/responses/internalServerError'
  /staff/{id}:
    get:
      summary: Get staff detail
      operationId: getStaffById
      tags:
        - staff
      parameters:
        - in: path
          name: id
          required: true
          description: Id of the staff
          schema:
            type: integer
      responses:
        '200':
          description: Successfully returned staff detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/staffItem'
                  - type: object
                    properties:
                      store:
                        type: object
                        properties:
                          id:
                            type: integer
                            format: number
                          phoneNumber:
                            type: string
                          address:
                            $ref: '#/components/schemas/address'
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '500':
          $ref: '#/components/responses/internalServerError'
    put:
      summary: Update staff
      operationId: updateStaff
      tags:
        - staff
      parameters:
        - in: path
          name: id
          required: true
          description: Id of the staff
          schema:
            type: integer
      requestBody:
        description: Payload of the staff to update
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
                active:
                  type: boolean
                phoneNumber:
                  type: string
                avatar:
                  type: string
                storeId:
                  type: integer
                  format: number
                address:
                  $ref: '#/components/schemas/address'
      responses:
        '204':
          description: Successfully updated staff
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '500':
          $ref: '#/components/responses/internalServerError'
    delete:
      summary: Delete staff
      operationId: deleteStaff
      tags:
        - staff
      parameters:
        - in: path
          name: id
          required: true
          description: The id of the staff
          schema:
            type: integer
      responses:
        '204':
          description: Successfully deleted staff
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
...
  /stores/{id}/staff:
    get:
      summary: Get staff employed in a store
      operationId: getStaffInStore
      tags:
        - store
      parameters:
        - in: path
          name: id
          required: true
          description: The id of the store
          schema:
            type: integer
      responses:
        '200':
          description: Successfully returned staff employed in a store
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          format: number
                        firstName:
                          type: string
                        lastName:
                          type: string
                        email:
                          type: string
                        addressId:
                          type: integer
                          format: number
                        storeId:
                          type: integer
                          format: number
                        active: true
                        phoneNumber:
                          type: string
                        avatar:
                          type: string
                        isStoreManager:
                          type: boolean
                        store:
                          type: object
                          properties:
                            id:
                              type: integer
                              format: number
                            storeManagerId:
                              type: integer
                              format: number
                            addressId:
                              type: integer
                              format: number
                            phoneNumber:
                              type: string
                  length:
                    type: integer
        '400':
          $ref: '#/components/responses/badRequest'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      name: auth_token
      in: cookie
  schemas:
...
    pagination:
      type: object
      properties:
        pageNumber:
          type: integer
          description: Current page number
        totalPages:
          type: integer
          description: Total pages available for pagination
        totalItems:
          type: integer
          description: Total items across all the pages
        itemsPerPage:
          type: integer
          description: Maximum number of items that the page can have
        next:
          type: string
          description: Link to the next page
        prev:
          type: string
          description: Link to the previous page
        first:
          type: string
          description: Link to the first page
        last:
          type: string
          description: Link to the last page
...
    address:
      type: object
      properties:
        id:
          type: integer
          format: number
          readOnly: true
        addressLine:
          type: string
        cityName:
          type: string
        stateName:
          type: string
        country:
          type: string
        postalCode:
          type: string
...
    staffItem:
      type: object
      properties:
        id:
          type: string
          format: number
          readOnly: true
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        isStoreManager:
          type: boolean
        active:
          type: boolean
        phoneNumber:
          type: string
        avatar:
          type: string
          format: uri
        address:
          $ref: '#/components/schemas/address'
...
    staffSummary:
      type: object
      properties:
        id:
          type: integer
          format: number
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        active:
          type: boolean
        phoneNumber:
          type: string
        avatar:
          type: string
          format: uri
        storeId:
          type: integer
          format: number
        isStoreManager:
          type: boolean
        address:
          $ref: '#/components/schemas/address'
...
  responses:
    badRequest:
      description: The request does not have the data that api needed
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    internalServerError:
      description: Server encountered unexpeted error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    unauthorized:
      description: Api key is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string
    notFound:
      description: Requested resource does not exist
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    forbidden:
      description: User do not have permission to perform the operation
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string